% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{split_sentences}
\alias{split_sentences}
\title{Splits a text into one or more sentences.}
\usage{
split_sentences(texts, locale = "")
}
\arguments{
\item{text}{A text.}
}
\value{
An array of sentences.
}
\description{
Splits a text into one or more sentences.
}
