% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{split_words}
\alias{split_words}
\title{Splits a text into one or more words.}
\usage{
split_words(
  texts,
  word_mask = 2147483647L,
  non_word_mask = 0L,
  lowercase = FALSE
)
}
\arguments{
\item{text}{A text.}
}
\value{
An array of words.
}
\description{
Splits a text into one or more words.
}
